yt/requirements.txt
asgiref==3.8.1
boto3==1.34.156
botocore==1.34.156
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
Django==4.1.7
ffmpeg-python==0.2.0
future==1.0.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
jmespath==1.0.1
moviepy==1.0.3
numpy==2.0.1
pillow==10.4.0
proglog==0.1.10
python-dateutil==2.9.0.post0
python-decouple==3.8
pytube @ git+https://github.com/pytube/pytube@a32fff39058a6f7e5e59ecd06a7467b71197ce35
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
tqdm==4.66.5
tzdata==2024.1
urllib3==2.2.2
